+++ x11/x11_icon.cpp
@@ -46,7 +46,7 @@ X11Icon::X11Icon(int width, int height, X11Widget *parent) : Icon(width, height)
 
 
 
-X11Icon::X11Icon(int width, int height, X11Widget *parent, const unsigned long *argb_data, unsigned long bg_color) : Icon(width, height),
+X11Icon::X11Icon(int width, int height, X11Widget *parent, const uint32 *argb_data, uint32 bg_color) : Icon(width, height),
     _pixmap(0),
     _mask(0)
 {
@@ -64,7 +64,7 @@ X11Icon::X11Icon(int width, int height, X11Widget *parent, const unsigned long *
 
         for(int y = 0; y< height; y++) {
             for(int x = 0; x < width; x++) {
-                unsigned long pixel = argb_data[(y *width) + x];
+                uint32 pixel = argb_data[(y *width) + x];
 
                 unsigned int alpha = (pixel & 0xFF000000) >> 24;
                 unsigned int red_in = (pixel & 0xFF0000) >> 16;
